完成触摸驱动
