+++ src/main.cpp
@@ -1,12 +1,14 @@
 #include <Arduino.h>
 #include "ui/ui.h"
 #include "uiport/lv_port_gfx_disp.h"
-#include "espWifiConfig.h"
+#include <Wire.h>
+
 #include "nvs_data_handle.h"
 #include "lv_disp_flush_Imp.h"
-#include "heartWeather.h" 
-#include <Wire.h>
-// #include "touch.h"
+
+#include "espWifiConfig.h"
+#include "heartWeather.h"
+
 
 #define DIRECT_MODE
 
@@ -17,14 +19,9 @@ const int reset_Pin = 0; // 设置重置按键引脚,用于删除WiFi信息
 const int wifi_LED = 2;  // 设置LED引脚
 uint8_t wifi_status = WL_IDLE_STATUS;
 
- 
 global_Time gl_time;
 global_Parameter global_Para;
 request_Result req_Result;
-/*******************
- *   需前置声明的函数
- ********************/
-void wificonnected(wl_status_t wl_status);
 
 /*******************
  *   全局类的声明
@@ -37,8 +34,9 @@ heartWeather myWeather("STm9u5f27O-X4vrvO");
 ui_handler_cb ui_wifi_hdl;
 
 /*******************
- *   定时器声明
+ *   需前置声明的函数
  ********************/
+
 
 /*******************
  *   setup开始
@@ -61,21 +59,21 @@ void setup()
   // 启动i2c串口
 
   Wire.begin();
-    Wire.setPins(8,9);
+  Wire.setPins(8, 9);
   Serial.begin(115200); // Set to a high rate for fast image transfer to a PC
 
   lv_port_gfx_Init();
-  Serial.print("init tftespidrv done!\n");
+  Serial.print("init gfxdrv done!\n");
+
+  
   ui_init();
   Serial.print("ui_init  done!\n");
   // 一切就绪, 启动LVGL任务
   xTaskNotifyGive(handleTaskLvgl);
   Serial.print("xTaskNotifyGive  done!\n");
 
-
-  //创建一个SNTP 校准时间的任务
-//Sntp_init();
-   
+  // 创建一个SNTP 校准时间的任务
+  // ntp_init();
 
   delay(1000);
 
@@ -118,47 +116,9 @@ void loop()
     }
   }
 
-  //dnsServer.processNextRequest();  // 检查客户端DNS请求
- // server.handleClient();           // 检查客户端(浏览器)http请求
-  //myWifiConfig.checkConnect(true); // 检测网络连接状态，参数true表示如果断开重新连接
+  // dnsServer.processNextRequest();  // 检查客户端DNS请求
+  // server.handleClient();           // 检查客户端(浏览器)http请求
+  // myWifiConfig.checkConnect(true); // 检测网络连接状态，参数true表示如果断开重新连接
 
   delay(30);
-}
-
-void wificonnected(wl_status_t wl_status)
-{
-  if (wifi_status != wl_status && wl_status == WL_CONNECTED)
-  {
-    wifi_status = wl_status;
-    //  保存变量
-    String ssid, pass;
-    myWifiConfig.getWifiInfo(ssid, pass);
-    global_Para.wifi_ssid = ssid;
-    global_Para.wifi_pass = pass;
-    savemyData(global_Para);
-
-    // 在welcome页面显示链接成功信息，
-    String lsWifi = "成功连接:(将在8秒后关闭此页面。)";
-    lsWifi = lsWifi + "\nWIFI:" + WiFi.SSID().c_str();
-    lsWifi = lsWifi + "\n闹钟IP:" + WiFi.localIP().toString().c_str();
-    _ui_label_set_property(ui_lbwifiInstr, 0, lsWifi.c_str());
-
-    Serial.print(lsWifi.c_str());
-    Serial.print('\n');
-
-    // 更换wifi状态icon
-    lv_obj_t *wifi_image = ui_comp_get_child(ui_panelTop1, 1);
-    if (wifi_image != NULL)
-    {
-      lv_img_set_src(wifi_image, &ui_img_png_wifi_full_png);
-    }
-    initTimer(); // start Timer
-    delay(8000);
-    _ui_screen_change(&ui_scToday, LV_SCR_LOAD_ANIM_NONE, 0, 0, NULL);
-    wifi_image = ui_comp_get_child(ui_panelTop2, 1);
-    if (wifi_image != NULL)
-    {
-      lv_img_set_src(wifi_image, &ui_img_png_wifi_full_png);
-    }
-  }
 }
