+++ src/uiport/lv_port_gfx_disp.h
@@ -13,7 +13,7 @@
  * ESP32 various dev board     : CS:  5, DC: 27, RST: 33, BL: 22, SCK: 18, MOSI: 23, MISO: nil
  * ESP32-C3 various dev board  : CS:  7, DC:  2, RST:  1, BL:  3, SCK:  4, MOSI:  6, MISO: nil
  * ESP32-S2 various dev board  : CS: 34, DC: 38, RST: 33, BL: 21, SCK: 36, MOSI: 35, MISO: nil
- * ESP32-S3 various dev board  : CS: 40, DC: 41, RST: 42, BL: 48, SCK: 36, MOSI: 35, MISO: nil
+ ** ESP32-S3 various dev board  : CS: 10, DC: 14, RST: -1, BL:  7, SCK: 12, MOSI: 11, MISO: 13
  * ESP8266 various dev board   : CS: 15, DC:  4, RST:  2, BL:  5, SCK: 14, MOSI: 13, MISO: 12
  * Raspberry Pi Pico dev board : CS: 17, DC: 27, RST: 26, BL: 28, SCK: 18, MOSI: 19, MISO: 16
  * RTL8720 BW16 old patch core : CS: 18, DC: 17, RST:  2, BL: 23, SCK: 19, MOSI: 21, MISO: 20
@@ -23,14 +23,23 @@
  * Seeeduino XIAO dev board    : CS:  3, DC:  2, RST:  1, BL:  0, SCK:  8, MOSI: 10, MISO:  9
  * Teensy 4.1 dev board        : CS: 39, DC: 41, RST: 40, BL: 22, SCK: 13, MOSI: 11, MISO: 12
  ******************************************************************************/
+
+
+
 #include <Arduino_GFX_Library.h>
+
+#include "touch_init.h"
 #include <lvgl.h>
 
 #define DIRECT_MODE // Uncomment to enable full frame buffer
 
+#define DISPLAY_ROTAION   3 
+
+
+ 
 extern TaskHandle_t handleTaskLvgl;
 void lv_port_gfx_Init();
 void lv_port_gfx_disp_init(Arduino_GFX* gfx);
-void Lvgl_gfx_Loop();
+
 #endif
