+++ src/uiport/touch_init.cpp
@@ -1,45 +1,13 @@
-/*******************************************************************************
- * Touch libraries:
- * XPT2046: https://github.com/PaulStoffregen/XPT2046_Touchscreen.git
- *
- * Capacitive touchscreen libraries
- * TouchLib: https://github.com/mmMicky/TouchLib.git
- ******************************************************************************/
-#include <XPT2046_Touchscreen.h>
+#include "touch_init.h"
 
-/* uncomment for XPT2046 */
-#ifndef _TOUCH_PAD_
-#define _TOUCH_PAD_
-#define TOUCH_XPT2046
-#define TOUCH_XPT2046_SCK 12
-#define TOUCH_XPT2046_MISO 13
-#define TOUCH_XPT2046_MOSI 11
-#define TOUCH_XPT2046_CS 10
-#define TOUCH_XPT2046_INT 18
-#define TOUCH_XPT2046_ROTATION 0
-#define TOUCH_XPT2046_SAMPLES 50
 
-// uncomment for most capacitive touchscreen
-// #define TOUCH_MODULES_FT5x06 // GT911 / CST_SELF / CST_MUTUAL / ZTW622 / L58 / FT3267 / FT5x06
-// #define TOUCH_MODULE_ADDR FT5x06_ADDR // CTS328_SLAVE_ADDRESS / L58_SLAVE_ADDRESS / CTS826_SLAVE_ADDRESS / CTS820_SLAVE_ADDRESS / CTS816S_SLAVE_ADDRESS / FT3267_SLAVE_ADDRESS / FT5x06_ADDR / GT911_SLAVE_ADDRESS1 / GT911_SLAVE_ADDRESS2 / ZTW622_SLAVE1_ADDRESS / ZTW622_SLAVE2_ADDRESS
-// #define TOUCH_SCL 5
-// #define TOUCH_SDA 6
-// #define TOUCH_RES -1
-// #define TOUCH_INT -1
-
-// Please fill below values from Arduino_GFX Example - TouchCalibration
 bool touch_swap_xy = false;
 
-int16_t touch_map_x1 = 229;
-int16_t touch_map_x2 = 3897;
-int16_t touch_map_y1 = 3970;
-int16_t touch_map_y2 = 333;
-/*
-int16_t touch_map_x1 = -1;
-int16_t touch_map_x2 = -1;
-int16_t touch_map_y1 = -1;
-int16_t touch_map_y2 = -1;
-*/
+int16_t touch_map_x1 = 272;
+int16_t touch_map_x2 = 3948;
+int16_t touch_map_y1 = 3938;
+int16_t touch_map_y2 = 378;
+ 
 int16_t touch_max_x = 0, touch_max_y = 0;
 int16_t touch_raw_x = 0, touch_raw_y = 0;
 int16_t touch_last_x = 0, touch_last_y = 0;
@@ -55,6 +23,7 @@ XPT2046_Touchscreen ts(TOUCH_XPT2046_CS, TOUCH_XPT2046_INT);
 TouchLib touch(Wire, TOUCH_SDA, TOUCH_SCL, TOUCH_MODULE_ADDR);
 
 #endif // TouchLib
+
 
 void touch_init(int16_t w, int16_t h, uint8_t r)
 {
@@ -197,5 +166,3 @@ bool touch_released()
   return false;
 #endif                           // TouchLib
 }
-
-#endif
